pyproject.toml
[project]
name = "aws-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "awslabs-aws-api-mcp-server>=0.2.9",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "mcp-agent>=0.1.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
